Rename muti_config to multi_config
